feat(EMS-900): DRY agenerateOTPAndUpdateAccount function
