+++ src/api/custom-resolvers/add-and-get-otp.ts
@@ -1,6 +1,6 @@
 import { Context } from '.keystone/types'; // eslint-disable-line
 import getAccountByField from '../helpers/get-account-by-field';
-import generate from '../helpers/generate-otp';
+import generateOTPAndUpdateAccount from '../helpers/generate-otp-and-update-account';
 import { AddOtpToAccountVariables } from '../types';
 
 /**
@@ -28,21 +28,8 @@ const addAndGetOtp = async (root: any, variables: AddOtpToAccountVariables, cont
       return { success: false };
     }
 
-    const otp = generate.otp();
-
-    const { securityCode, salt, hash, expiry } = otp;
-
-    // update the account.
-    const accountUpdate = {
-      otpSalt: salt,
-      otpHash: hash,
-      otpExpiry: expiry,
-    };
-
-    await context.db.Exporter.updateOne({
-      where: { id: exporter.id },
-      data: accountUpdate,
-    });
+    // generate OTP and update the account
+    const { securityCode } = await generateOTPAndUpdateAccount(context, exporter.id);
 
     return {
       success: true,
